seperated storage apis logic
